fix: adapting tests for github CI

+++ src/test/config/global-setup.ts
@@ -28,9 +28,13 @@ before(async () => {
     .resolve(import.meta.dirname, "compose.test.yml")
     .replace(/\\/g, "/")
 
-  await $`docker compose -f "${composeFileAbsolutePath}" -p gamerchallengestest up -d`
-  //   await new Promise((r) => setTimeout(r, 1000))
-  await waitForPostgres()
+  if (!process.env.CI) {
+    await $`docker compose -f "${composeFileAbsolutePath}" -p gamerchallengestest up -d`
+    //   await new Promise((r) => setTimeout(r, 1000))
+    await waitForPostgres()
+  } else {
+    console.log("Running in CI, skipping docker compose up")
+  }
 
   const prismaSchemaAbsolutePath = path.resolve(
     import.meta.dirname,
@@ -47,7 +51,9 @@ beforeEach(async () => {
 after(async () => {
   server.close()
   await prisma.$disconnect()
-  await $`docker compose -p gamerchallengestest down > /dev/null 2>&1`
+  if (!process.env.CI) {
+    await $`docker compose -p gamerchallengestest down > /dev/null 2>&1`
+  }
 })
 
 async function truncateTables() {
